Send coverage results to Code Climate
Add one job to the Circle CI configuration, where `rake coverage` is run between
the `./cc-test-reporter` `before-build` and `after-build` commands.

The `codeclimate-test-reporter` gem is not needed anymore, and we can use a more
recent version of simplecov, too.

I also changed `spec_helper.rb` so that SimpleCov is only run when the the
`rake coverage` task is invoked.